+++ src/click/model/ipv4-l3-click-protocol.cc
@@ -29,8 +29,6 @@ namespace ns3
 
 NS_LOG_COMPONENT_DEFINE("Ipv4L3ClickProtocol");
 
-const uint16_t Ipv4L3ClickProtocol::PROT_NUMBER = 0x0800;
-
 NS_OBJECT_ENSURE_REGISTERED(Ipv4L3ClickProtocol);
 
 TypeId