+++ src/internet/model/arp-l3-protocol.cc
@@ -28,8 +28,6 @@ namespace ns3
 
 NS_LOG_COMPONENT_DEFINE("ArpL3Protocol");
 
-const uint16_t ArpL3Protocol::PROT_NUMBER = 0x0806;
-
 NS_OBJECT_ENSURE_REGISTERED(ArpL3Protocol);
 
 TypeId