+++ src/internet/model/ipv6-l3-protocol.h
@@ -14,6 +14,7 @@
 #include "ipv6-routing-protocol.h"
 #include "ipv6.h"
 
+#include "ns3/iana-ieee802-numbers.h"
 #include "ns3/ipv6-address.h"
 #include "ns3/net-device.h"
 #include "ns3/traced-callback.h"
@@ -59,9 +60,9 @@ class Ipv6L3Protocol : public Ipv6
     static TypeId GetTypeId();
 
     /**
-     * @brief The protocol number for IPv6 (0x86DD).
+     * @brief The protocol number for IPv6
      */
-    static constexpr uint16_t PROT_NUMBER = 0x86DD;
+    static constexpr uint16_t PROT_NUMBER = iana::Ieee802Numbers::IPV6;
 
     /**
      * @enum DropReason