+++ src/network/utils/radiotap-header.h
@@ -344,8 +344,8 @@ class RadiotapHeader : public Header
         HE_DATA1_STBC_KNOWN = 0x0200,         /**< STBC known */
         HE_DATA1_SPTL_REUSE_KNOWN =
             0x0400, /**< Spatial Reuse known (Spatial Reuse 1 for HE TRIG PPDU format) */
-        HE_DATA1_SPTL_REUSE2_KNOWN = 0x0800, /**< Spatial Reuse 2 known (HE TRIG PPDU format),
-                                                STA-ID known (HE MU PPDU format) */
+        HE_DATA1_SPTL_REUSE2_KNOWN = 0x0800, /**< Spatial Reuse 2 known (HE TRIG
+                                                PPDU format), STA-ID known (HE MU PPDU format) */
         HE_DATA1_SPTL_REUSE3_KNOWN = 0x1000, /**< Spatial Reuse 3 known (HE TRIG PPDU format) */
         HE_DATA1_SPTL_REUSE4_KNOWN = 0x2000, /**< Spatial Reuse 4 known (HE TRIG PPDU format) */
         HE_DATA1_BW_RU_ALLOC_KNOWN = 0x4000, /**< data BW/RU allocation known */