+++ src/point-to-point/model/ppp-header.cc
@@ -9,6 +9,7 @@
 #include "ns3/abort.h"
 #include "ns3/assert.h"
 #include "ns3/header.h"
+#include "ns3/iana-ppp-numbers.h"
 #include "ns3/log.h"
 
 #include <iostream>
@@ -51,14 +52,14 @@ PppHeader::Print(std::ostream& os) const
 
     switch (m_protocol)
     {
-    case 0x0021: /* IPv4 */
+    case iana::PppDllNumbers::IPV4_DLL:
         proto = "IP (0x0021)";
         break;
-    case 0x0057: /* IPv6 */
+    case iana::PppDllNumbers::IPV6_DLL:
         proto = "IPv6 (0x0057)";
         break;
     default:
-        NS_ASSERT_MSG(false, "PPP Protocol number not defined!");
+        NS_ASSERT_MSG(false, "PPP DLL Protocol number not defined!");
     }
     os << "Point-to-Point Protocol: " << proto;
 }